+++ CHANGELOG.md
@@ -16,6 +16,7 @@ CHANGELOG
 * @OwenEdwards improved handling of deprecated use of Button component ([view](https://github.com/videojs/video.js/pull/3236))
 * @forbesjo added chrome for PR tests ([view](https://github.com/videojs/video.js/pull/3235))
 * @MCGallaspy added vttjs to the self-hosting guide ([view](https://github.com/videojs/video.js/pull/3229))
+* @chrisauclair added ARIA region and label to player element ([view](https://github.com/videojs/video.js/pull/3227))
 
 --------------------